SW-5122 Move the current home view to TerrawareHomeView and stub out the ParticipantHomeView
