+++ src/scenes/Home/ParticipantHomeView.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ParticipantHomeView = () => {
+  return <p>Participant Experience</p>;
+};
+
+export default ParticipantHomeView;
